refactor: moved logic code into hooks file
